updates
- change int64 to the default int
- printStats logic changed
- use len function to count bytes instead of incrementing in loop
- moved newline check inside isSpace block

+++ wc_test.go
@@ -7,26 +7,28 @@ import (
 var stats = GetStats("test.txt")
 
 func TestNumLines(t *testing.T) {
-	want := int64(7137)
+	want := 7137
 	if stats.numLines != want {
 		t.Fatalf("want %d, got %d", want, stats.numLines)
 	}
 }
 
 func TestNumWords(t *testing.T) {
-	want := int64(58159)
+	want := 58159
 	if stats.numWords != want {
 		t.Fatalf("want %d, got %d", want, stats.numWords)
 	}
 }
+
 func TestNumChars(t *testing.T) {
-	want := int64(339120)
+	want := 339120
 	if stats.numChars != want {
 		t.Fatalf("want %d, got %d", want, stats.numChars)
 	}
 }
+
 func TestNumBytes(t *testing.T) {
-	want := int64(341836)
+	want := 341836
 	if stats.numBytes != want {
 		t.Fatalf("want %d, got %d", want, stats.numBytes)
 	}
